automagik/cli/requirements.txt
click>=8.0.0
python-dotenv>=1.0.0
sqlalchemy>=1.4.0
tabulate>=0.8.0
httpx>=0.24.0
inquirer>=3.1.3
celery>=5.3.0
redis>=4.5.0
pytest>=7.0.0 